Update bumpskin.py
I discovered that I used a wrong thermal conductivity in the roof calculations. The numbers for the roof matches the numbers of the wall since they have both the same thickness, thermal conductivity and heat capacity. I changed the number for the wall for layer lev =3 from 0.005 to 0.006 because we use the no_flux option. Has not a significant effect on other numbers